backend/requirements.txt
pyserial
requests
fastapi
uvicorn
opencv-python
numpy
python-multipart
